+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
 import 'package:flutter_osm_plugin_example/src/home/main_example.dart';
 import 'package:flutter_osm_plugin_example/src/search_example.dart';
 import 'package:flutter_osm_plugin_example/src/simple_example_hook.dart';
@@ -25,9 +26,29 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData.light(useMaterial3: true),
-      initialRoute: "/home",
+      initialRoute: "/simple",
       routes: {
         "/home": (context) => const MainPageExample(),
+        "/simple": (context) => Scaffold(
+              appBar: AppBar(
+                title: const Text('simple'),
+              ),
+              body: OSMViewer(
+                controller: SimpleMapController(
+                  initPosition: GeoPoint(
+                    latitude: 47.4358055,
+                    longitude: 8.4737324,
+                  ),
+                  markerHome: const MarkerIcon(
+                    icon: Icon(Icons.home),
+                  ),
+                ),
+                zoomOption: const ZoomOption(
+                  initZoom: 16,
+                  minZoomLevel: 11,
+                ),
+              ),
+            ),
         "/old-home": (context) => const OldMainExample(),
         "/hook": (context) => const SimpleHookExample(),
         //"/adv-home": (ctx) => AdvandedMainExample(),